feat(cache): array handler
handler array handler

+++ src/main/java/org/fz/nettyx/serializer/struct/basic/c/signed/Clong4.java
@@ -27,7 +27,7 @@ public class Clong4 extends Cint {
      * @param value the length
      */
     public Clong4(Object value) {
-        super((Integer) value);
+        super(value);
     }
 
     /**
